Moved all optimal sample methods into shared Strategy class
