+++ SALib/sample/morris.py
@@ -27,7 +27,7 @@ from __future__ import division
 import random as rd
 import numpy as np
 
-from SALib.sample.morris_strategies import (SampleMorris)
+from SALib.sample.morris_strategies import SampleMorris
 from SALib.sample.optimal_trajectories import GlobalOptimisation
 from SALib.sample.morris_strategies.local import LocalOptimisation
 from SALib.sample.morris_strategies.brute import BruteForce
