+++ SALib/sample/morris_util.py
@@ -5,9 +5,6 @@ Helper functions for Morris trajectories
 
 from __future__ import division
 
-from scipy.spatial.distance import cdist
-from itertools import combinations, islice
-
 import numpy as np
 import random as rd
 
@@ -104,150 +101,3 @@ def generate_x_star(k, num_levels, grid_step):
 
 def compute_delta(num_levels):
     return float(num_levels) / (2 * (num_levels - 1))
-
-
-def check_input_sample(input_sample, num_params, N):
-    '''
-    Checks input sample is:
-        - the correct size
-        - values between 0 and 1
-    '''
-    assert type(input_sample) == np.ndarray, \
-        "Input sample is not an numpy array"
-    assert input_sample.shape[0] == (num_params + 1) * N, \
-        "Input sample does not match number of parameters or groups"
-    assert np.any((input_sample >= 0) | (input_sample <= 1)), \
-        "Input sample must be scaled between 0 and 1"
-
-
-def compute_distance(m, l):
-    '''
-    Computes the distance between two trajectories
-    '''
-
-    if np.shape(m) != np.shape(l):
-        raise ValueError("Input matrices are different sizes")
-    if np.array_equal(m, l):
-        print("Trajectory %s and %s are equal" % (m, l))
-        distance = 0
-    else:
-        distance = np.array(np.sum(cdist(m, l)), dtype=np.float32)
-
-    return distance
-
-
-def compute_distance_matrix(input_sample, N, num_params, groups=None,
-                            local_optimization=False):
-    """Computes the distance between every trajectory
-
-    Arguments
-    ---------
-    input_sample : numpy.ndarray
-        The input sample of trajectories for which to compute
-        the distance matrix
-    N : int
-        The number of samples
-    num_params : int
-        The number of factors
-    groups : tuple, default=None
-        the mapping between factors and groups
-    local_optimization : bool, default=False
-
-    Returns
-    -------
-    distance_matrix : numpy.ndarray
-
-    """
-    num_groups = None
-    if groups:
-        num_groups = groups[0].shape[1]
-        check_input_sample(input_sample, num_groups, N)
-    else:
-        check_input_sample(input_sample, num_params, N)
-
-    index_list = make_index_list(N, num_params, num_groups)
-    distance_matrix = np.zeros((N, N), dtype=np.float32)
-
-    for j in range(N):
-        input_1 = input_sample[index_list[j]]
-        for k in range(j + 1, N):
-            input_2 = input_sample[index_list[k]]
-
-            if local_optimization is True:
-                distance_matrix[j, k] = compute_distance(input_1, input_2)
-
-            distance_matrix[k, j] = compute_distance(input_1, input_2)
-    return distance_matrix
-
-
-def find_maximum(scores, N, k_choices):
-    """Finds the `k_choices` maximum scores from `scores`
-
-    Arguments
-    ---------
-    scores : numpy.ndarray
-    N : int
-    k_choices : int
-    """
-    if not isinstance(scores, np.ndarray):
-        raise TypeError("Scores input is not a numpy array")
-
-    index_of_maximum = int(scores.argmax())
-    maximum_combo = nth(combinations(
-        list(range(N)), k_choices), index_of_maximum, None)
-    return sorted(maximum_combo)
-
-
-def take(n, iterable):
-    "Return first n items of the iterable as a list"
-    return list(islice(iterable, n))
-
-
-def nth(iterable, n, default=None):
-    """Returns the nth item or a default value
-
-    Arguments
-    ---------
-    iterable : iterable
-    n : int
-    default : default=None
-        The default value to return
-    """
-
-    if type(n) != int:
-        raise TypeError("n is not an integer")
-
-    return next(islice(iterable, n, None), default)
-
-
-def make_index_list(N, num_params, groups=None):
-    """
-
-    Returns
-    -------
-    list of numpy.ndarray
-    """
-    if groups is None:
-        groups = num_params
-
-    index_list = []
-    for j in range(N):
-        index_list.append(np.arange(groups + 1) + j * (groups + 1))
-    return index_list
-
-
-def compile_output(input_sample, N, num_params, maximum_combo, groups=None):
-    """
-    """
-
-    if groups is None:
-        groups = num_params
-
-    check_input_sample(input_sample, groups, N)
-
-    index_list = make_index_list(N, num_params, groups)
-
-    output = np.zeros((np.size(maximum_combo) * (groups + 1), num_params))
-    for counter, x in enumerate(maximum_combo):
-        output[index_list[counter]] = np.array(input_sample[index_list[x]])
-    return output
